update many index related code
fix some bugs (e.g. stale index node not GC'ed correctly)

optimize block BSI space cost

+++ pkg/vm/engine/aoe/storage/db/sched/flushblkidx.go
@@ -87,7 +87,7 @@ func (e *flushBlockIndexEvent) Execute() error {
 			panic("logic error")
 		}
 		nodes = append(nodes, vec.MNode)
-		startPos := int(schema.BlockMaxRows * uint64(meta.Idx))
+		startPos := schema.BlockMaxRows * uint64(meta.Idx)
 		bsi, err := index.BuildBitSlicedIndex([]*vector.Vector{&vec.Vector}, schema.ColDefs[colIdx].Type, int16(colIdx), startPos)
 		if err != nil {
 			panic(err)
